requirements.txt
apipkg==1.5
appdirs==1.4.3
argon2-cffi==19.1.0
attrs==19.3.0
authenticator==1.1.3
badge==2019.4.13
beautifulsoup4==4.8.1
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.1
Chameleon==3.6.2
chardet==3.0.4
check-manifest==0.40
Click==7.0
cmarkgfm==0.4.2
coinbase==2.1.0
cryptography==2.8
defusedxml==0.6.0
devpi==2.2.0
devpi-client==5.0.0
devpi-common==3.4.0
devpi-server==5.2.0
devpi-web==4.0.0
Django==2.2.7
django-render-block==0.6
docutils==0.15.2
execnet==1.7.1
filelock==3.0.12
format-keys==2019.4.13
github-name==2019.4.13
github-remote==2019.4.14
google-auth==1.7.1
hupper==1.9
idna==2.8
importlib-metadata==0.23
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
markdown-table==2019.4.13
MarkupSafe==1.1.1
MasterMVC==1.7.3
mkdir==2019.4.13
more-itertools==7.2.0
mysql-connector-python==8.0.18
orderdict==2019.9.25
packaging==19.2
passlib==1.7.1
PasteDeploy==2.0.1
paystack==1.5.0
Pillow==6.2.1
pkginfo==1.5.0.1
plaster==1.0
plaster-pastedeploy==0.7
pluggy==0.13.0
protobuf==3.10.0
public==2019.4.13
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycryptodome==3.9.0
Pygments==2.4.2
pyparsing==2.4.2
pyramid==1.10.4
pyramid-chameleon==0.3
python-dateutil==2.8.0
python-readme-generator==2019.4.13
pytz==2019.3
readme-badges==2019.4.13
readme-docstring==2019.4.13
readme-generator==2019.4.13
readme-renderer==24.0
repoze.lru==0.7
requests==2.22.0
requests-toolbelt==0.9.1
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
setupcfg==2019.4.13
shields==2019.4.13
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
strictyaml==1.0.5
toml==0.10.0
tox==3.14.0
tqdm==4.36.1
translationstring==1.3
twine==2.0.0
urllib3==1.25.6
values==2019.4.13
venusian==1.2.0
virtualenv==16.7.7
waitress==1.3.1
webencodings==0.5.1
WebOb==1.8.5
Whoosh==2.7.4
write==2019.4.13
zipp==0.6.0
zope.deprecation==4.4.0
zope.interface==4.6.0
